LayoutTests/inspector/debugger/file-system-project-mapping-expected.txt

Tests file system project mappings.


Running: testAutomaticMapping
Adding file system.
Adding network resource.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> 
    filesystem:/var/www/bar.js -> 
Adding mapping between network and file system resources.
Emulate reloading inspector.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> http://localhost/html/foo.js
    filesystem:/var/www/bar.js -> http://localhost/bar.js
Removing mapping between network and file system resources.
Emulate reloading inspector.
UISourceCode uri to url mappings:
    filesystem:/var/www/html/foo.js -> 
    filesystem:/var/www/bar.js -> 

Running: testScriptFileOnReloadWithDirtyFile
Adding file system.
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content>
    script: <foo content>
  //@ sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Editing uiSourceCode:
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Committing uiSourceCode with live edit failure:
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited>
    script: <foo content>
  //@ sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Reloading with saved but diverged uiSourceCode:
Loading script.
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited>
    script: <foo content>
  //@ sourceURL=http://localhost/html/foo.js  
 
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Editing uiSourceCode again:
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
Committing uiSourceCode again (with live edit success now):
Dumping uiSourceCode and script content:
    uiSourceCode: <foo content edited again>
    script: <foo content edited again>
 //@ sourceURL=http://localhost/html/foo.js
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Reloading page:
Loading script.
 - hasDivergedFromVM: false
 - sourceFrame._muted: false
Editing uiSourceCode again and reloading while it is dirty:
Loading script.
 - hasDivergedFromVM: true
 - sourceFrame._muted: true
